services/user-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
requests==2.31.0
python-multipart==0.0.6
itsdangerous==2.1.2

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.12.0

# Authentication
passlib[bcrypt]>=1.7.4
pydantic[email]>=2.0.0
PyJWT>=2.8.0
